+++ HISTORY.md
@@ -81,6 +81,7 @@
         * new "svg poly" export format for vector pen trails, optimized for speed and laser-cutting
         * export vector pen trails as embroidery files (DST, EXP)
         * added pen trails export options to the project ("file") menu
+        * added ability to switch octaves in the piano keyboard menu, thanks, ego-lay-atman-bay!
 * **Notable Changes:**
     1. Blocks all the way
         * renamed "primitive" blocks into "extension"
@@ -122,6 +123,9 @@
     * updated extensions.md
 * **Translation Updates:**
     * German
+
+### 2024-07-24
+* widgets: added ability to switch octaves in the piano keyboard widget, thanks, ego-lay-atman-bay!
 
 ### 2024-07-22
 * edited help screen for doDeleteBlock, converted to PNG, reduced file size